+++ src/MyMediaLiteExperimental/AttrToFactor/BPRMF_ItemMapping_Optimal.cs
@@ -1,4 +1,4 @@
-// Copyright (C) 2010, 2011 Zeno Gantner
+// Copyright (C) 2010, 2011, 2012 Zeno Gantner
 //
 // This file is part of MyMediaLite.
 //
@@ -45,7 +45,7 @@ namespace MyMediaLite.AttrToFactor
 		///
 		public override void LearnAttributeToFactorMapping()
 		{
-			this.factor_bias = new double[num_factors];
+			this.factor_bias = new float[num_factors];
 
 			for (int i = 0; i < num_factors; i++)
 			{
@@ -53,13 +53,13 @@ namespace MyMediaLite.AttrToFactor
 				Console.Error.WriteLine("fb {0}: {1}", i, factor_bias[i]);
 			}
 
-			this.attribute_to_factor = new Matrix<double>(NumItemAttributes, num_factors);
+			this.attribute_to_factor = new Matrix<float>(NumItemAttributes, num_factors);
 			MatrixExtensions.InitNormal(attribute_to_factor, InitMean, InitStdDev);
 
 			for (int i = 0; i < num_iter_mapping; i++)
 				IterateMapping();
 
-			_MapToLatentFactorSpace = Utils.Memoize<int, double[]>(__MapToLatentFactorSpace);
+			_MapToLatentFactorSpace = Utils.Memoize<int, float[]>(__MapToLatentFactorSpace);
 		}
 
 		///
@@ -99,23 +99,23 @@ namespace MyMediaLite.AttrToFactor
 				// update attribute-factor parameter for latent factors which are different between the items
 				foreach (int a in attr_i_over_j)
 				{
-					double m_af = attribute_to_factor[a, f];
+					float m_af = attribute_to_factor[a, f];
 					double update = w_uf / (1 + Math.Exp(x_uij)) - reg_mapping * m_af;
-					attribute_to_factor[a, f] = m_af + learn_rate_mapping * update;
+					attribute_to_factor[a, f] = (float) (m_af + learn_rate_mapping * update);
 				}
 				foreach (int a in attr_j_over_i)
 				{
-					double m_af = attribute_to_factor[a, f];
+					float m_af = attribute_to_factor[a, f];
 					double update = -w_uf / (1 + Math.Exp(x_uij)) - reg_mapping * m_af;
-					attribute_to_factor[a, f] = m_af + learn_rate_mapping * update;
+					attribute_to_factor[a, f] = (float) (m_af + learn_rate_mapping * update);
 				}
 			}
 		}
 
 		///
-		protected override double[] __MapToLatentFactorSpace(int item_id)
+		protected override float[] __MapToLatentFactorSpace(int item_id)
 		{
-			var factor_representation = new double[num_factors];
+			var factor_representation = new float[num_factors];
 
 			for (int j = 0; j < num_factors; j++)
 				factor_representation[j] = factor_bias[j];
@@ -128,7 +128,7 @@ namespace MyMediaLite.AttrToFactor
 		}
 
 		///
-		protected override double[] MapToLatentFactorSpace(int item_id)
+		protected override float[] MapToLatentFactorSpace(int item_id)
 		{
 			return _MapToLatentFactorSpace(item_id);
 		}
@@ -139,8 +139,7 @@ namespace MyMediaLite.AttrToFactor
 			return string.Format(
 				CultureInfo.InvariantCulture,
 				"{0} num_factors={1} reg_u={2} reg_i={3} reg_j={4} num_iter={5} learn_rate={6} reg_mapping={7} num_iter_mapping={8} learn_rate_mapping={9} init_mean={10} init_stddev={11}",
-				this.GetType().Name, num_factors, reg_u, reg_i, reg_j, NumIter, learn_rate, reg_mapping, num_iter_mapping, learn_rate_mapping, InitMean, InitStdDev
-			);
+				this.GetType().Name, num_factors, reg_u, reg_i, reg_j, NumIter, learn_rate, reg_mapping, num_iter_mapping, learn_rate_mapping, InitMean, InitStdDev);
 		}
 	}
 }
